refactored the store into modules
